requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
pydantic==1.10.12
motor==3.1.1
python-dateutil==2.8.2
